third_party/WebKit/LayoutTests/geolocation-api/not-enough-arguments.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests how Geolocation handles receiving too few arguments.");

shouldThrow("window.navigator.geolocation.getCurrentPosition()");
shouldThrow("window.navigator.geolocation.watchPosition()");
shouldThrow("window.navigator.geolocation.clearWatch()", '"TypeError: Failed to execute \'clearWatch\' on \'Geolocation\': 1 argument required, but only 0 present."');
</script>
</body>
</html>
