third_party/WebKit/LayoutTests/geolocation-api/not-enough-arguments-expected.txt

Tests how Geolocation handles receiving too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.navigator.geolocation.getCurrentPosition() threw exception TypeError: Failed to execute 'getCurrentPosition' on 'Geolocation': 1 argument required, but only 0 present..
PASS window.navigator.geolocation.watchPosition() threw exception TypeError: Failed to execute 'watchPosition' on 'Geolocation': 1 argument required, but only 0 present..
PASS window.navigator.geolocation.clearWatch() threw exception TypeError: Failed to execute 'clearWatch' on 'Geolocation': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
